+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 - Added basic define macro support
+- Removed unsupported matrix constructors
 
 ## 1.0.0 (2026-04-07)
 - Added Korean, German translations